fix(phan-quyen): van_phong_ids dùng bigint khớp van_phong_id + REVOKE anon helper (#154)
van_phong_id là bigint nên van_phong_ids và can_access_van_phong(target) phải
bigint (bản integer cũ fail 42883 lúc apply — prod đã apply bản bigint qua MCP).
Sync file migration + DOWN script (drop signature bigint) cho khớp prod.
Thêm REVOKE EXECUTE anon/PUBLIC trên 4 helper RLS (đã apply prod) — khớp
hardening 20260601, silence advisor 0028.

+++ docs/db-rollback/20260609_van_phong_hard_scope_DOWN.sql
@@ -27,8 +27,8 @@ BEGIN
   END LOOP;
 END $$;
 
--- 2. Drop helper functions.
-DROP FUNCTION IF EXISTS public.can_access_van_phong(int);
+-- 2. Drop helper functions. (signature bigint — khớp van_phong_id kiểu bigint)
+DROP FUNCTION IF EXISTS public.can_access_van_phong(bigint);
 DROP FUNCTION IF EXISTS public.current_user_vp_scope();
 DROP FUNCTION IF EXISTS public.current_user_is_accounting();
 DROP FUNCTION IF EXISTS public.current_user_cross_vp();
